requirements.txt
datasets
tqdm
mmh3
langchain-community
langchain-google-genai
pinecone>=5.0.0
langchain-huggingface
sentence-transformers
chromadb
flask
python-dotenv
twilio
gunicorn
requests
beautifulsoup4
trafilatura
langchain-text-splitters

